# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "app para cotizaciones de criptos",
        "title": "Cripto Api",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
...
        "/cotizaciones": {
            "post": {
                "description": "Save a new quotation",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "cryptocurrencies"
                ],
                "summary": "Save a quotation",
                "parameters": [
                    {
                        "description": "Cotizacion to be saved",
                        "name": "cotizacion",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/primerProjecto_internal_entities_criptomonedas.Cotizacion"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response with a message",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    }
                }
            }
        },
...
        "/cryptocurrencies/lastcotization/{nombre}": {
            "get": {
                "description": "Get the latest quote for a given cryptocurrency name",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "cryptocurrencies"
                ],
                "summary": "Get latest quote by cryptocurrency name",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Cryptocurrency name",
                        "name": "nombre",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response with the latest quote",
                        "schema": {
                            "$ref": "#/definitions/primerProjecto_internal_entities_criptomonedas.Cotizacion"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    }
                }
            }
        },
...
        "/usuarios/{id}/cotizaciones": {
            "get": {
                "description": "Find all cryptocurrencies by filter for a specific user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "cryptocurrencies"
                ],
                "summary": "Find all cryptocurrencies by filter",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "User ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Nombre",
                        "name": "nombre",
                        "in": "query"
                    },
                    {
                        "type": "number",
                        "description": "Minimum Cotizacion",
                        "name": "min_cotizacion",
                        "in": "query"
                    },
                    {
                        "type": "number",
                        "description": "Maximum Cotizacion",
                        "name": "max_cotizacion",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Start Date in RFC3339 format",
                        "name": "start_date",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "End Date in RFC3339 format",
                        "name": "end_date",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page Size",
                        "name": "page_size",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Page Number",
                        "name": "page_number",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response with summary and data",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "gin.H": {
            "type": "object",
            "additionalProperties": {}
        },
        "primerProjecto_internal_entities_criptomonedas.Cotizacion": {
            "description": "Estructura que define una cotización de criptomoneda.",
            "type": "object",
            "properties": {
                "cotizacion": {
                    "description": "Cotizacion es el valor de la criptomoneda en un momento específico.\n@example 50000.00",
                    "type": "number"
                },
                "cripto_id": {
                    "description": "CriptoMoneda_ID es el identificador de la criptomoneda asociada.\n@example 1",
                    "type": "integer"
                },
                "fecha": {
                    "description": "Fecha es la fecha y hora en que se registró la cotización.\n@example 2024-07-29T12:00:00Z",
                    "type": "string"
                },
                "id": {
                    "description": "ID es el identificador único de la cotización.\n@example 123",
                    "type": "integer"
                },
                "manual": {
                    "description": "Manual indica si la cotización fue ingresada manualmente.\n@example true",
                    "type": "boolean"
                },
                "usuario_id": {
                    "description": "UsuarioId es el identificador del usuario que ingresó la cotización.\n@example 42",
                    "type": "integer"
                }
            }
        },
...
    }
}